+++ src/FoodTracker.Domain/Shared/IMacroSource.cs
@@ -3,5 +3,5 @@ namespace FoodTracker.Domain.Shared;
 public interface IMacroSource
 {
     ServingUnit ServingUnit { get; }
-    MacroSnapshot ComputeMacros(decimal quantity);
+    MacroSnapshot ComputeMacros(double quantity);
 }
